refactor(bridge): extract secret-encrypted edit handling
